Refactor: Firebase config & api
