+++ index.html
@@ -180,7 +180,8 @@
     <script src="./js/UI/employeeTable.js" type="module"></script>
     <script src="./js/UI/modal.js" type="module"></script>
     <script src="./js/UI/skills.js" type="module"></script>
-    <script src="./js/firebase/firebase.js" type="module"></script>
+    <script src="./js/firebase/config.js" type="module"></script>
+    <script src="./js/firebase/api.js" type="module"></script>
     <script src="./js/filter.js" type="module"></script>
     <script src="./js/toggleTheme.js" type="module"></script>
   </body>
