+++ js/firebase/api.js
@@ -1,6 +1,4 @@
-import { initializeApp } from "https://www.gstatic.com/firebasejs/10.4.0/firebase-app.js";
 import {
-  getDatabase,
   ref,
   onValue,
   update,
@@ -10,30 +8,15 @@ import {
 } from "https://www.gstatic.com/firebasejs/10.4.0/firebase-database.js";
 
 import {
-  getStorage,
   uploadBytes,
   getDownloadURL,
   ref as strRef,
 } from "https://www.gstatic.com/firebasejs/10.4.0/firebase-storage.js";
 
+import { db, storage } from "./config.js";
+
 import { showToast } from "../utils/toast.js";
 import { closeModal } from "../UI/modal.js";
-
-const firebaseConfig = {
-  apiKey: "AIzaSyCVJaFfoEBM3-ZqT89iRzQ-1K9dKIoThO8",
-  authDomain: "hrm-app-6cb10.firebaseapp.com",
-  databaseURL:
-    "https://hrm-app-6cb10-default-rtdb.asia-southeast1.firebasedatabase.app",
-  projectId: "hrm-app-6cb10",
-  storageBucket: "hrm-app-6cb10.appspot.com",
-  messagingSenderId: "588841693832",
-  appId: "1:588841693832:web:7a98a8759551243ee0653d",
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const db = getDatabase(app);
-const storage = getStorage(app);
 
 let loadingState = false;
 export const isLoading = () => loadingState;
